third_party/WebKit/LayoutTests/media/controls-cast-button-narrow.html

<!DOCTYPE html>
<title>media controls cast button</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="media-file.js"></script>
<script src="media-controls.js"></script>
<video width="100" height="200" controls></video>
<script>
async_test(function(t) {
    var video = document.querySelector("video");
    video.src = findMediaFile("video", "content/test");

    video.onloadedmetadata = t.step_func_done(function() {
        // Pretend we have a cast device
        internals.mediaPlayerRemoteRouteAvailabilityChanged(video, true);

        // Video should not have a cast button since the video is too narrow
        var button = castButton(video);
        assert_equals(button.style.display, "none", "button should not be visible");

        // It should not have an overlay cast button too. Instead, the button
        // should appear in the overflow menu. Tested somewhere else.
        button = overlayCastButton(video);
        assert_equals(button.style.display, "none", "button should not be visible");
    });

    function castButton(element) {
        var controlID = "-internal-media-controls-cast-button";
        var button = mediaControlsElement(internals.shadowRoot(element).firstChild, controlID);
        if (!button)
            throw "Failed to find cast button";
        return button;
    }

    function overlayCastButton(element) {
        var controlID = "-internal-media-controls-overlay-cast-button";
        return mediaControlsElement(internals.shadowRoot(element).firstChild, controlID);
        return button
        if (!button)
            throw "Failed to find cast button";
        return button;
    }
});
</script>
